fix(web): render a file/suite node's own stdout/stderr in the viewer
The row model treated a node as either a container (expandable children)
or a diagnostics carrier, so once a file's top-level stdout/stderr shared
the file node with its tests, that output was dropped. Let a container
carry its own diagnostics too, toggled by the Details badge independently
of expanding children (mirroring the live reporter, whose TreeNode
already renders both). Extract the pure row-model logic to a testable
module and unit-test the regression.
